Remove pymongo as runtime and buildtime dependency
pymongo is used only in optional mongodb driver. Note that mongodb was
retired from RDO long time ago so i think we can remove pymongo as
default runtime dependency.

Also, pymongo is not needed at build time as we are not running unit
tests and we are not building html docs for zaqar.

Change-Id: I81f2ee804b889e9c29a7dda906cde438fce48db4